--- forklift_driver/.vscode/settings.json
@@ -1,32 +1,0 @@
-{
-    "files.associations": {
-        "ostream": "cpp",
-        "cmath": "cpp",
-        "iostream": "cpp",
-        "string": "cpp",
-        "array": "cpp",
-        "tuple": "cpp",
-        "utility": "cpp",
-        "random": "cpp",
-        "condition_variable": "cpp",
-        "iomanip": "cpp"
-    },
-    "python.autoComplete.extraPaths": [
-        "/home/ericlai/common_ws/devel/lib/python3/dist-packages",
-        "/home/ericlai/gpm_project/devel/lib/python3/dist-packages",
-        "/home/ericlai/catkin_ws/devel/lib/python3/dist-packages",
-        "/home/ericlai/test_ws/devel/lib/python3/dist-packages",
-        "/home/ericlai/3dlidar_ws/devel/lib/python3/dist-packages",
-        "/home/ericlai/visual_ws/devel/lib/python3/dist-packages",
-        "/opt/ros/noetic/lib/python3/dist-packages"
-    ],
-    "python.analysis.extraPaths": [
-        "/home/ericlai/common_ws/devel/lib/python3/dist-packages",
-        "/home/ericlai/gpm_project/devel/lib/python3/dist-packages",
-        "/home/ericlai/catkin_ws/devel/lib/python3/dist-packages",
-        "/home/ericlai/test_ws/devel/lib/python3/dist-packages",
-        "/home/ericlai/3dlidar_ws/devel/lib/python3/dist-packages",
-        "/home/ericlai/visual_ws/devel/lib/python3/dist-packages",
-        "/opt/ros/noetic/lib/python3/dist-packages"
-    ]
-}
