--- forklift_driver/include/hello.h
@@ -1,5 +1,0 @@
-#ifndef _HELLO_H
-#define _HELLO_H
-void hello();
-
-#endif
