--- forklift_msg/CMakeLists.txt
@@ -1,30 +1,0 @@
-cmake_minimum_required(VERSION 3.0.2)
-project(forklift_msg)
-
-find_package(catkin REQUIRED COMPONENTS
-  roscpp
-  rospy
-  std_msgs
-  message_generation
-)
-
-add_message_files(
-  FILES
-  meteorcar.msg
-)
-
-generate_messages(
-  DEPENDENCIES
-  std_msgs
-)
-
-catkin_package(
-  CATKIN_DEPENDS 
-  message_runtime
-  std_msgs
-)
-
-include_directories(
-  include
-  ${catkin_INCLUDE_DIRS}
-)
