--- forklift_msg/package.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>forklift_msg</name>
-  <version>0.0.0</version>
-  <description>The fork_msg package</description>
-
-  <maintainer email="user@todo.todo">user</maintainer>
-
-  <license>TODO</license>
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>roscpp</build_depend>
-  <build_export_depend>roscpp</build_export_depend>
-  <exec_depend>roscpp</exec_depend>
-
-  <build_depend>rospy</build_depend>
-  <exec_depend>rospy</exec_depend>   
-
-  <build_depend>std_msgs </build_depend>
-  <build_export_depend>std_msgs</build_export_depend>
-  <exec_depend>std_msgs</exec_depend>
-
-  <build_depend>actionlib_msgs</build_depend>
-  <build_export_depend>actionlib_msgs</build_export_depend> 
-  <exec_depend>actionlib_msgs</exec_depend>
-
-  <build_depend>message_generation</build_depend>
-  <build_export_depend>message_runtime</build_export_depend>
-  <exec_depend>message_runtime</exec_depend>
-
-  <export>
-  </export>
-</package>
